adding useChipFlavour comp
